Restructure files and make requiring common resources easier
