+++ core/modules/config.js
